requirements.txt
langchain==0.1.0
langchain-google-genai==0.0.5
langchain-community==0.0.10
chromadb==0.4.22
sentence-transformers==2.3.1
huggingface-hub==0.22.2
mysql-connector-python==8.2.0
pymysql==1.1.0
SQLAlchemy==2.0.23
python-dotenv==1.0.0
streamlit==1.29.0
pandas==2.1.4
numpy==1.26.3
PyPDF2==3.0.1
python-multipart==0.0.6
requests==2.31.0

